Fixed includes to be more clear and removed unnecessary files
