+++ tensorflow/lite/micro/kernels/ifx_fast/conv/conv_ops_ref.h
@@ -23,6 +23,18 @@ namespace ops {
 namespace micro {
 namespace conv {
 
+inline PaddingType RuntimePaddingType(TfLitePadding padding) {
+    switch (padding) {
+        case TfLitePadding::kTfLitePaddingSame:
+            return PaddingType::kSame;
+        case TfLitePadding::kTfLitePaddingValid:
+            return PaddingType::kValid;
+        case TfLitePadding::kTfLitePaddingUnknown:
+        default:
+            return PaddingType::kNone;
+    }
+}
+
 TfLiteStatus EvalConvUInt8Reference(TfLiteConvParams* params, OpData* data,
                                     const TfLiteEvalTensor* input,
                                     const TfLiteEvalTensor* filter,
